First Commit with Push to GitHub
This is an early stage push. The majority of the Administration area will be unusable until I finish porting from 1.2 to
1.3 but that should be completed soon. Taking a holiday away shortly but I'll be back to complete it and then I can
complete the process of improving the overall usability of certain functions.

Main goals after the port include stripping out obsoleted Javascript and replacing with handrolled jQuery plugins reused
from Dux and then prepping a default style for initial install.

After this the final goal is an installer. Big big priority